[Renderer] Change SdfRenderer_Params to SdfRendererProperties files
