+++ src/Renderer/SdfRendererProperties.hpp
@@ -1,0 +1,37 @@
+#pragma once
+
+#include "Core/AttributeListModel.hpp"
+#include "Core/FloatAttribute.hpp"
+
+#include <QObject>
+
+struct SdfRenderer_Params
+{
+    float absorptionCoefficient = 0.5f;
+    float absorptionCutoff = 0.25f;
+    float marchMultiplier = 1.8f;
+    int maxVolumeMarchSteps = 20;
+    int maxVolumeLightMarchSteps = 4;
+    int maxSdfSphereSteps = 15;
+    int maxOpaqueShadowMarchSteps = 10;
+};
+
+class SdfRendererProperties : public QObject
+{
+    Q_OBJECT
+    Q_PROPERTY(AttributeListModel *attributes MEMBER _attributes NOTIFY
+                   attributesChanged)
+  public:
+    SdfRendererProperties(QObject *parent = nullptr);
+    ~SdfRendererProperties() = default;
+
+    /// \brief Create a SdfRenderer_Params,
+    /// used for the Rendering thread.
+    const SdfRenderer_Params sdfRenderer_Params() const;
+
+  public:
+    Q_SIGNAL void attributesChanged();
+
+  private:
+    AttributeListModel *_attributes;
+};
